requirements.txt
pandas==2.1.4
pyyaml==6.0.1
jinja2==3.1.2
requests==2.31.0
aiohttp==3.9.1
psutil==5.9.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1 